requirements.txt
Authlib==0.15.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-datastore==2.1.3
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
